core: Fix mismatched type

+++ include/Serial.h
@@ -3,6 +3,6 @@
 #include <HardwareSerial.h>
 
 extern "C" void print__i__(int n);
-extern "C" void print__f__(int n);
+extern "C" void print__f__(float n);
 extern "C" void print__s__(char c[]);
 extern "C" void begin(int b);
